you can delete or edit a post only if you created it
on front end and back end as well

+++ backend/routes/posts.js
@@ -64,11 +64,17 @@ router.put(
       _id: req.body.id,
       title: req.body.title,
       content: req.body.content,
-      imagePath: imagePath
+      imagePath: imagePath,
+      creator: req.userData.userId
     });
     console.log(post);
-    Post.updateOne({ _id: req.params.id }, post).then(result => {
-      res.status(200).json({ message: "Update successful!" });
+
+    Post.updateOne({ _id: req.params.id, creator: req.userData.userId }, post).then(result => {
+      if(result.nModified > 0){
+        res.status(200).json({ message: "Update successful!" });
+      }else{
+        res.status(401).json({ message: "Not authorized!" });
+      }
     });
   }
 );
@@ -108,10 +114,15 @@ router.get('/:id', (req, res, next) => {
   });
 });
 
-router.delete('/:id',(req, res, next) => {
-  Post.deleteOne({_id: req.params.id}).then(result => {
+router.delete('/:id', checkAuth, (req, res, next) => {
+  Post.deleteOne({_id: req.params.id, creator: req.userData.userId}).then(result => {
      console.log(result);
-     res.status(200).json({message: "Post Deleted!"});
+     if(result.n > 0){
+      res.status(200).json({message: "Post Deleted!"});
+     }else{
+      res.status(401).json({message: "Not authorized!"});
+
+     }
 
   });
 });
